+++ src/EngineManaged/Bindings/UniformBuffer.cpp
@@ -102,10 +102,10 @@ void Flood::UniformBuffer::SetUniform(System::String^ slot, System::Collections:
     for each(Flood::Color _element in vec)
     {
         auto _marshal0 = ::Color();
-        _marshal0.r = _element.R;
-        _marshal0.g = _element.G;
-        _marshal0.b = _element.B;
-        _marshal0.a = _element.A;
+        _marshal0.r = (byte)(uint8)_element.R;
+        _marshal0.g = (byte)(uint8)_element.G;
+        _marshal0.b = (byte)(uint8)_element.B;
+        _marshal0.a = (byte)(uint8)_element.A;
         auto _marshalElement = _marshal0;
         _tmpvec.push_back(_marshalElement);
     }
